+++ Ambar/Model/DTO/RateDTO.cs
@@ -14,7 +14,7 @@ namespace Ambar.Model.DTO
         private decimal surplus_level;
         private string service;
         private int year;
-        private short month;
+        private int month;
 
         public Guid Rate_ID { get => rate_id; set => rate_id = value; }
         public decimal Basic_Level { get => basic_level; set => basic_level = value; }
@@ -22,6 +22,6 @@ namespace Ambar.Model.DTO
         public decimal Surplus_Level { get => surplus_level; set => surplus_level = value; }
         public string Service { get => service; set => service = value; }
         public int Year { get => year; set => year = value; }
-        public short Month { get => month; set => month = value; }
+        public int Month { get => month; set => month = value; }
     }
 }
