+++ Ambar/Model/DTO/RatesReportDTO.cs
@@ -12,7 +12,7 @@ namespace Ambar.ViewController
         [CsvHelper.Configuration.Attributes.Name("Anio")]
         private int year;
         [CsvHelper.Configuration.Attributes.Name("Mes")]
-        private short month;
+        private int month;
         [CsvHelper.Configuration.Attributes.Name("Tarifa Basica")]
         private decimal basic_level;
         [CsvHelper.Configuration.Attributes.Name("Tarifa Intermedia")]
@@ -23,7 +23,7 @@ namespace Ambar.ViewController
         [CsvHelper.Configuration.Attributes.Name("Anio")]
         public int Year { get => year; set => year = value; }
         [CsvHelper.Configuration.Attributes.Name("Mes")]
-        public short Month { get => month; set => month = value; }
+        public int Month { get => month; set => month = value; }
         [CsvHelper.Configuration.Attributes.Name("Tarifa Basica")]
         public decimal Basic_Level { get => basic_level; set => basic_level = value; }
         [CsvHelper.Configuration.Attributes.Name("Tarifa Intermedia")]
